Moved Course to a separate module and moved subcomponents too and imported Course into App
