+++ 02. Communicating With Server/courseinfo/src/App.js
@@ -1,48 +1,5 @@
-const Header = (props) => {
-    return (
-        <h1>{props.course.name}</h1>
-    );
-};
-
-const Content = (props) => {
-    return (
-        <div>
-            {props.course.parts.map(part =>
-                <Part key={part.id} part={part} />
-            )}
-        </div>
-    );
-};
-
-const Part = (props) => {
-    return (
-        <p>
-            {props.part.name} {props.part.exercises}
-        </p>
-    );
-};
-
-const Total = (props) => {
-    const initialValue = 0;
-    const sumWithInitialValue = props.course.parts.reduce(
-        (previousValue, currentValue) => previousValue + currentValue.exercises,
-        initialValue
-    );
-
-    return (
-        <p><strong>Total of {sumWithInitialValue}</strong></p>
-    );
-};
-
-const Course = (props) => {
-    return (
-        <div>
-            <Header course={props.course} />
-            <Content course={props.course} />
-            <Total course={props.course} />
-        </div>
-    );
-};
+import React from "react";
+import Course from "./components/Course";
 
 const App = () => {
     const courses = [
